Move api/server/router/session to daemon/server/router/session
